requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
pandas
requests
PyPDF2
chromadb